GroupBy using hash maps for `AVG()` expressions (#1150)
* First simple optimization

* Allow any aggregate function

* text

This implements an initial implementation of GROUP BY that uses a hash map to aggregate the groups and therefore doesn't require a sorted input. Currently this implementation is only used when all the expressions in a group by have the form AVG(subexpresssionThatDoesn'tContainAggregates), but this commit already contains much of the infrastructure that is needed to also implement this version of GROUP BY also for the other aggregates and for more complex expressions.