Refactor index builder, i.p. makes it easier to change permutation build order (#1163)
Code from the large `createFromFile` method was moved to separate methods, which makes the code much clearer. In particular, it is now easy to change the order in which the permutations are built. On the side, fix the problem introduced in #1154 which led to some blocks of the permutations being too large.