Add watchdog for CancellationHandle (#1170)
Implement a mechanism that detects if we check for a cancellation request too sparsely (currently less than every 50ms). This mechanism can be enabled or disabled at compile time (for maximal efficiency) as well as at runtime (to not clutter the log e.g. of unit tests).
The goal is to use this mechanism to automatically find all the places where checks are missing.