Add a function for adding thousans separators into strings (#1182)
This function finds all substrings that represent numbers in a larger string (using a regex) and formats them nicely using thousands separators. This function can e.g. be used to nicely format a JSON output for printing with readable numbers (note that this makes the string invalid JSON because the thousands separators are not allowed there).