Allow disabling cancellation checks (#1194)
The checks whether a query shall be cancelled can now be disabled at compile time. This makes it possible to analyze the performance impact of these checks. Additionally it is possible to only enable the cancellation, but not the watchdog that checks whether the cancellation checks are performed frequently enough.