Clean up several concepts (#1214)
There are now concepts `SimilarToAny` and `SameAsAny` in the `util/TypeTraits.h` header that check whether a type is the same as or similar to one of a list of other types.  Previously there only was a version of `SimilarToAny` which had a name that was not as clear. Similarly we now have `SameAsAnyTypeIn` and `SimilarToAnyTypeIn` which check if a type is the same as or similar to any of the types contained in a template (e.g. if a type is one of the types contained in a certain `std::variant<...>` or `std::tuple<..>`.