Use the patterns from the permutations and no longer load `ql:has-pattern` into RAM (#1223)
PRs #1168 and #1177 have added the subject patterns as two additional columns to the OSP&OPS and PSO&POS permutations. PR #1226 has added the triples of the `ql:has-pattern` predicate to the PSO&POS permutations. Now use this information instead of the old patterns, which did cost a lot of RAM. We tried a few queries involving patterns and the speed is very similar to that of the previous implementation.

NOTE: This is an index-breaking change. The old `.index.patterns` file stored the `ql:has-pattern` predicate (for each subject its pattern) and the information which pattern consists of which predicates. Now the `.index.patterns` file only stores the latter information. The file size therefore is significantly reduced and no longer depends on the size of the dataset (but only on how many distinct patterns there are, typically few). For example, for Wikidata, the file size reduced from 17 GB to 2.8 GB. For UniProt, the reduction is from 152 GB (which does not fit into the RAM of our standard machines) to something very small (because UniProt is very regular and there are only very few distinct patterns).