Give each index a (reasonably) unique ID (#984)
The index ID is currently just a concatenation of the index name, and the number of triples, subjects, predicates, and objects. This makes it unlikely (but not impossible) that two different indexes have the same ID. The ID can be obtained via the API with `cmd=get-index-id`. 

This feature is useful for applications like https://github.com/ad-freiburg/qlever-petrimaps, which have an internal cache that depends on the index. When the index changes, the application should be able to notice that, and with the index ID, it now can.