Add a helper class that blocks until a given `shared_ptr` is destroyed (#1233)
Introduce `ad_utility::PointerGuard`, a class that stores a `shared_ptr` and is able to give out copies of this `shared_ptr`. The destructor blocks until the `PointerGuard` is the single owner of the `shared_ptr` which is equivalent to blocking until the pointee is destroyed alongside with the `PointerGuard`. This class can be used to nest the lifetime of objects even in asynchronous settings where shared ownership via `shared_ptr`s is used. We currently use it inside the `QueryHub` class to make sure that neither the `QueryHub` nor the underlying `io_context` are accessed after they are destroyed. While this was not the source of the not so spurious failures in the `HttpTest` we have still decided to integrate this tool to make the reasoning about the destruction of these objects in test scenarios easier.