Implement `isIRI`, `isBlank`, `isLiteral`, `isNumeric`, `bound` (#1184)
A reasonably efficient implementation that precomputes the ranges of IRIs, blank nodes, and literals in the internal and external vocabulary, respectively (building on #1235). At function evaluation time, it is then checked per `Id` whether it contains a vocabulary index and whether that index is contained in one of the two ranges. For items in the local vocabulary, the string is looked up. For `isNumeric`, we can simply check the datatype bits of the `Id` (there are currently only two numeric datatypes in QLever: `int` and `double`). For `bound` we simply check whether the `Id` is `UNDEF` or not.

Addresses #1183 . Here is an example query on Wikidata that evaluates all functions on a cross-section of entities: https://qlever.cs.uni-freiburg.de/wikidata/06dsWT . Here is a query that evaluates`isBlank` for a large number of entities: https://qlever.cs.uni-freiburg.de/wikidata/9Srzzy .