Add more cancellation checks (#1193)
This PR adds more cancellation checks to allow the application to react to cancellation requests much faster. 
In particular, this commit adds checks for the complete Expression module and the operations that use expressions (FILTER, BIND, and GROUP BY). It also adds checks to the JOIN operation. Note that OPTIONAL and multi-column joins currently are not yet cancelable.