Write external vocabulary directly to `.vocabulary.external` (#1253)
So far, the code that merges the partial vocabularies (class `VocabularyMerger`) first wrote the external vocabulary to a temporary file which was afterward copied to the final `. vocabulary.external` file. This wasted space during the index build and on HDD also cost considerable time. Now the result of the merge is written directly to `. vocabulary.external`.