Efficient GROUP BY with COUNT (#1263)
Queries of the following kind have taken very long so far or even failed when the underlying index scan has many triples:

```
SELECT ?object (COUNT(?object) AS ?count) WHERE {
  ?subject rdf:type ?object
} GROUP BY ?object
```

The reason is that all blocks of the index scan were read and decompressed, and for the duration of the query, all triples were kept in RAM and then iterated over.

With this change, queries of the form above are now processed by iterating over the metadata of the blocks of the underlying index scan. When all triples of a block have the same `col1Id`, we can tell that from the metadata alone and we get all the information for our result from the metadata. Only when there is more than one `col1Id` do we need to load and decompress the block and iterate over its triples. This yields a very large speed-up when the number of groups is small relative to the total number of triples. In particular, this now allows interactive object AC queries on UniProt.