Add cancellation checks for query export (#1272)
Also enable the cancellation (manually or via timeout) of a query when the query result is already present in the ID space, but the cancellation happens during the export of the query result. Note that for the lazy streaming export (currently TSV, CSV, Turtle) a partial result has already been sent to the user and there is no good way to report the error to the user in this case. We currently force a hard close of the connection, to make the user aware that the result was not complete, but we have yet to discuss more graceful methods.