Store `col0Id` in permutations, smaller `.meta` files, improved full scan (#1289)
The `col0Id` is now stored in each permutation (one of the `.index.???` files) just like the `col1Id` and the `col2Id`. The `.meta` file for each permutation now only stores information for those `col0Id` that span multiple blocks. Despite the increased capabilities (see below), the relevant code is now simpler and easier to understand because `col0Id`, `col1Id`, and `col2Id` can now be treated in an analogous fashion.

This change reduces the size of the `.meta` files to a small fraction of their previous size (they used to be very large for permutations with many "relations", that is, for SPO and SOP with many subjects, and for OSP and OPS with many objects). In contrast, the size of the `.index.???` files inceases only little because they are compressed and due to their sort order contain long runs of triples with the same `col0Id`. In total, this change reduces the total size of the index files by half or even more.

In a closely related change, full index scans (queries involving patterns like `?s ?p ?o` and which need to read the full triples with `col0Id`, `col1Id`, `col2Id` from a permutation), can now be treated just like normal scans. This removes a lot of the extra code that accumulated over time for the special treatment of full scans. Also, these scans are now as efficient (per block) as a normal index scan. Another nice side effect is that full scans with GROUP BY and COUNT can now be processed with the code from #1263.