Fix the address and thread sanitizer builds on newer kernels (#1304)
Recent versions of the linux kernel and also recent versions of the Ubuntu distributions have increased the number of bits that are used for virtual address layout randomization. The sanitizer tools have to be aware of this randomization but have not yet picked up that change. Our current fix is to manually reduce the number of randomization bits back to 28 (the previous value) for our GitHub actions builds.