Make the tests for the HTTP server fast again. (#1302)
While a query waits to be assigned to a computation thread, there is a timer that polls the cancellation handle every 25 ms. Previously there was no way to cancel this timer, which meant that each HTTP request needed at least 25 ms to complete. This was especially annoying in the unit tests which run many requests in tight loops to trigger potential lifetime and threading issues. This is now fixed as this timer can be canceled and is cancelled as early as possible.