Add generic `ProgressBar` class + use for index build log  (#1315)
The new `ProgressBar` class works for any computation that proceeds in countable steps (for example, the lines of an input file or the triples of a permutation). The final number of steps does not need to be known in advance. After each batch of steps, an update string is produced with information about the number of steps processed  so far and statistics about the average processing speed, the speed for the last batch, and the fastest and slowest speed for a batch so far. The batch size is purely a parameter of the class, the actual computation does not need to proceed in batches in any way.

This is now used in all places in the index build, where we showed progress so far, namely: parsing input triples, merging words from the partial vocabularies, converting triples from local to global IDs, and sorting the triples for the various permutation pairs. The index log thus becomes more compact and more informative at the same time.

On the side, make a few other improvements to the index log. For example, when a permutation pair is built, now says which permutations these are (e.g., SPO and SOP) already at the beginning and not only when it's done. This requires a corresponding change in the `index-stats` command of the `qlever` script.