Implement Transitive Paths using binary search instead of a hash map (#1313)
The previous implementation of transitive paths first converted the subtree (the graph which is traversed) into a `HashMap<HashSet<Id>>`  that represented the adjacency list of the graph. This took a lot of time for large graphs even if the actual transitive path computation was cheap because only a few starting nodes were considered.
We now use the sorted subtree as a graph directly using binary search. This is much faster especially since the subtree is often already presorted as it is a single `IndexScan`. 

Currently both implementations are present in the Code. The active implementation can be switched by using a runtime parameter. The old implementation is kept around mostly for benchmarking purposes, as the new implementation should always be cheaper.