Disable timing tests for both Docker and macOS builds (#1325)
The Docker build that performs ARM-64 cross compilation has unreliable timing when run as part of a GitHub workflow. We therefore now disable timing tests not only for the macOS build (like before), but also for the Docker builds. That way the publication of Docker images is not prevented by spuriously failing timing tests anymore.