Fix a subtle bug in `CompressedExternalIdTableSorter` (#1328)
Under certain (rare) conditions, `CompressedExternalIdTableSorter::sortedBlocks` produced empty blocks, which led to an exception. This is now fixed. Also make sure that an exception in the building of a permutation pair does not result in a deadlock anymore (as it did for the bug just mentioned), but in a proper error message.