Reintroduce the `TEXTLIMIT` feature (#1245)
This commit reintroduces the `TEXTLIMIT` feature that allows to limit the number of results from the fulltext index per mentioned entity. (The semantics are described in the CIKM 17 paper in detail). The `TEXTLIMIT` is now implemented as a separate operation which makes it more visible and easier to reason about in comparison to the old monolithic design of the text operations.