Implement functions  `STRDT()` and `STRLANG()` (#1360)
These construct a string literal from a given literal + datatype (`STRDT`) or a string literal from a given literal + language tag (`STRLANG)`.
Note: Currently numbers and dates etc. are not correctly handled: If the datatype for STRDT is for example `xsd:int`, then the result should be an ID that directly encodes the number. This will be fixed in the future, as the function, which can handle these conversions automatically, is part of another PR.