Implement functions `LANG()` and `LANGMATCHES()` (#1382)
These are currently implement trivially by retrieving the full string of each literal and then checking the substring which stores the language tag. We plan a much more efficient implementation in the future that stores the language tag directly in the ID.
Currently a lot of expressions including LANG() functions and relational expressions like `==` or `<=` don't work as expected because of restrictions in those expressions which will be addressed in the near future.