Proper handling of (all kinds of) empty queries (#1383)
Empty queries were properly handled so far (before the parsing), but queries that consisted only of whitespace, comments, or prefix declarations were not. The first two produced a segmentation fault, the last produced a confusing error message. Now none of these queries lead to a segmentation fault anymore, and in each case, the same proper error message is returned.