Correctly interleave the internal and external vocabulary (#1405)
So far, operations that require a lexicographic ordering of string literals or IRIs (for example: ORDER BY, certain FILTERs, prefix searches) were usually not correct when items from both the internal and the external vocabulary were involved. Specifically, so far all words from the external vocabulary were handled as if they were lexicographically larger than all words from the internal vocabulary. This is now fixed. An item is now searched in the combined vocabularies by first binary-searching it in the internal vocabulary and then, if it is not found there, completing the search in the external vocabulary. To make this more efficient, it is ensured that at least each 1000th item (this number is currently a hard-coded constant) from the total order is contained in the internal vocabulary. Note that unlike before, the IDs of the items are now no longer consecutive in the internal vocabulary (but still in the external vocabulary). Also, both vocabularies are now compressed in the same way and there is one shared codebook.

IMPORTANT: This is an index-breaking change. In particular, the `.vocabulary...`  files have now different names.