Fix query planning for join with `SERVICE` clause (#1422)
There is a bug in the optimization introduced by #1341, which checks whether the sibling operand of a join with a `SERVICE` operations has a small result (and if yes, passes that result to the `SERVICE` clause as `VALUES` in order to reduce the result size of the `SERVICE` call). Instead of just adding the sibling operand to the existing `Service` operation (which led to an object that should have been `const` being changed), now a new `Service` operation that includes the sibling is created. In particular, this fixes https://qlever.cs.uni-freiburg.de/osm-planet, which worked before #1341 and now works again as it should.