Avoid no-op generator wrapper when limit is no-op (#1477)
Currently the `Result` class wraps lazy results in a generator if the corresponding `Operation` does not support limits natively in order to trim the result sizes in post processing. If the limit is empty however this generator wrapper is completely redundant. This commit removes this extra generator if neither a LIMIT nor an OFFSET exists. This possibly makes it easier to debug such queries, the runtime overhead should be negligible if the lazy blocks are large enough.